feat: store para dialog e correções do dialog

+++ frontend/src/pages/LessonList.vue
@@ -15,63 +15,80 @@
         :category="lesson.category"
         :description="lesson.description"
         :onEdit="goToEditLessonPage"
-        :onDelete="deleteCard"
+        :onDelete="() => deleteCard(lesson.id)"
         :onClick="goToViewPage"
         class="lessons__card"
       />
     </div>
+
+    <QaDialog />
   </q-page>
 </template>
 
 <script setup>
 import { Notify } from 'quasar';
+import { useDialogStore } from 'src/stores/dialogStore';
 import { useLessonStore } from 'src/stores/lessonStore';
-import { computed, onMounted } from 'vue';
+import { computed, onMounted, watch } from 'vue';
 import { useRouter } from 'vue-router';
 import CardComponent from '../components/CardComponent.vue';
+import QaDialog from '../components/QaDialog.vue';
 
 defineOptions({
   name: 'LessonList'
 });
 
-const lessonStore = useLessonStore();
-const lessons = computed(() => lessonStore.lessons);
+  const lessonStore = useLessonStore();
+  const lessons = computed(() => lessonStore.lessons);
 
-onMounted(() => {
-  lessonStore.getLessons();
-});
+  onMounted(() => {
+    lessonStore.getLessons();
+  });
 
-const router = useRouter();
+  const router = useRouter();
 
-const goToEditLessonPage = (id) => {
-  router.push(`/lessons/${id}/edit`);
-};
+  const goToEditLessonPage = (id) => {
+    router.push(`/lessons/${id}/edit`);
+  };
 
-const goToViewPage = (id) => {
-  router.push(`/lessons/${id}/view`);
-}
+  const goToViewPage = (id) => {
+    router.push(`/lessons/${id}/view`);
+  }
 
-const goToCreatePage = () => {
-  router.push(`/lessons/new`);
-}
+  const goToCreatePage = () => {
+    router.push(`/lessons/new`);
+  }
 
-const deleteCard = async (id) => {
-    try {
-      await lessonStore.deleteLesson(id);
+  const dialogStore = useDialogStore();
 
-      Notify.create({
-        type: 'positive',
-        message: 'Lição deletada com sucesso!',
-        position: 'top-right'
-      });
-    } catch(error) {
-      Notify.create({
-        type: 'negative',
-        message: 'Não foi possível deletar a lição!',
-        position: 'top-right'
-      });
-    }
-}
+  const deleteCard = (id) => {
+    dialogStore.openDialog();
+
+    watch(
+      () => dialogStore.confirmed,
+      async (confirmed) => {
+        if (confirmed) {
+          try {
+            await lessonStore.deleteLesson(id);
+
+            Notify.create({
+              type: 'positive',
+              message: 'Lição deletada com sucesso!',
+              position: 'top-right'
+            });
+          } catch (error) {
+            console.log(error);
+            Notify.create({
+              type: 'negative',
+              message: 'Não foi possível deletar a lição!',
+              position: 'top-right'
+            });
+          }
+        }
+      },
+      { immediate: false }
+    );
+  }
 </script>
 
 <style lang="scss" scoped>
